+++ core/core/models/__init__.py
@@ -1,7 +1,5 @@
-# from .experiment import Experiment
+from .experiment import Experiment
 from .feature_set import FeatureSet
-
-# from .model import Model
 from .project import Project
 from .stream_processor import StreamProcessor
 from .message_bus import MessageBus
