+++ core/core/models/feature_set.py
@@ -10,4 +10,9 @@ class FeatureSet(sqlm.SQLModel, table=True):  # type: ignore[call-arg]
     key_field: str
     ts_field: str
 
-    project_name: str
+    project_name: str = sqlm.Field(default=None, foreign_key="project.name")
+    project: "Project" = sqlm.Relationship(  # type: ignore[name-defined]
+        sa_relationship_kwargs={"uselist": False}
+    )
+
+    experiments: list["Experiment"] = sqlm.Relationship(back_populates="feature_set")  # type: ignore[name-defined]
